refactor(api-headless-cms-storage): move registries to features/ convention
Restructure the four filter/sorting DI registries (FieldFilterPathRegistry,
FieldFilterValueTransformRegistry, FieldFilterCreateRegistry,
FieldSortingRegistry) from separate abstractions/ and implementations/
directories into per-feature directories under src/features/, matching the
convention used by entryTableManager and syncTableManager elsewhere in the
codebase. Each feature directory now holds its own abstractions.ts and
registry implementation file. Updated all internal imports across filtering,
handlers, and the package's public index.

Also regenerated api-headless-cms-sql's tsconfig files, which dropped an
unused @webiny/plugins project reference.
