+++ packages/api-headless-cms-sql/tsconfig.json
@@ -8,7 +8,6 @@
     { "path": "../db-utils" },
     { "path": "../feature" },
     { "path": "../handler" },
-    { "path": "../plugins" },
     { "path": "../utils" }
   ],
   "compilerOptions": {
@@ -30,8 +29,6 @@
       "@webiny/feature": ["../feature/src"],
       "@webiny/handler/*": ["../handler/src/*"],
       "@webiny/handler": ["../handler/src"],
-      "@webiny/plugins/*": ["../plugins/src/*"],
-      "@webiny/plugins": ["../plugins/src"],
       "@webiny/utils/*": ["../utils/src/*"],
       "@webiny/utils": ["../utils/src"]
     }
