+++ packages/api-headless-cms-storage/src/features/FilterRegistriesFeature.ts
@@ -1,12 +1,12 @@
 import { createFeature } from "@webiny/feature/api/index.js";
-import { FieldFilterPathRegistry } from "../abstractions/FieldFilterPathRegistry.js";
-import { FieldFilterValueTransformRegistry } from "../abstractions/FieldFilterValueTransformRegistry.js";
-import { FieldFilterCreateRegistry } from "../abstractions/FieldFilterCreateRegistry.js";
-import { FieldSortingRegistry } from "../abstractions/FieldSortingRegistry.js";
-import { FieldFilterPathRegistryImpl } from "../implementations/FieldFilterPathRegistryImpl.js";
-import { FieldFilterValueTransformRegistryImpl } from "../implementations/FieldFilterValueTransformRegistryImpl.js";
-import { FieldFilterCreateRegistryImpl } from "../implementations/FieldFilterCreateRegistryImpl.js";
-import { FieldSortingRegistryImpl } from "../implementations/FieldSortingRegistryImpl.js";
+import { FieldFilterPathRegistry } from "./fieldFilterPath/abstractions.js";
+import { FieldFilterValueTransformRegistry } from "./fieldFilterValueTransform/abstractions.js";
+import { FieldFilterCreateRegistry } from "./fieldFilterCreate/abstractions.js";
+import { FieldSortingRegistry } from "./fieldSorting/abstractions.js";
+import { FieldFilterPathRegistryImpl } from "./fieldFilterPath/FieldFilterPathRegistry.js";
+import { FieldFilterValueTransformRegistryImpl } from "./fieldFilterValueTransform/FieldFilterValueTransformRegistry.js";
+import { FieldFilterCreateRegistryImpl } from "./fieldFilterCreate/FieldFilterCreateRegistry.js";
+import { FieldSortingRegistryImpl } from "./fieldSorting/FieldSortingRegistry.js";
 import { createPlainObjectPathHandler } from "../handlers/plainObjectPathHandler.js";
 import { createLocationFolderIdPathHandler } from "../handlers/locationFolderIdPathHandler.js";
 import { createDatetimeTransformHandler } from "../handlers/datetimeTransformHandler.js";
