+++ packages/api-headless-cms-storage/src/index.ts
@@ -35,8 +35,8 @@ export type { Field } from "./filtering/fields/types.js";
 export type { FilterItemFromStorage } from "./filtering/fields/types.js";
 
 /* DI registries. */
-export { FieldFilterPathRegistry } from "./abstractions/FieldFilterPathRegistry.js";
-export { FieldFilterValueTransformRegistry } from "./abstractions/FieldFilterValueTransformRegistry.js";
-export { FieldFilterCreateRegistry } from "./abstractions/FieldFilterCreateRegistry.js";
-export { FieldSortingRegistry } from "./abstractions/FieldSortingRegistry.js";
+export { FieldFilterPathRegistry } from "./features/fieldFilterPath/abstractions.js";
+export { FieldFilterValueTransformRegistry } from "./features/fieldFilterValueTransform/abstractions.js";
+export { FieldFilterCreateRegistry } from "./features/fieldFilterCreate/abstractions.js";
+export { FieldSortingRegistry } from "./features/fieldSorting/abstractions.js";
 export { FilterRegistriesFeature } from "./features/FilterRegistriesFeature.js";
